+++ cli/tests/test_buckets.rs
@@ -9,13 +9,7 @@ fn test_bucket_lifecycle() {
     let new_bucket_name = format!("{}/test-source-{}", owner, Uuid::new_v4());
 
     // Create bucket
-    let output = cli.run([
-        "create",
-        "bucket",
-        &new_bucket_name,
-        "--transform-tag",
-        "generic_simple.0.QR5PW4VZ",
-    ]);
+    let output = cli.run(["create", "bucket", &new_bucket_name]);
     assert!(output.contains(&new_bucket_name), "{}", output);
 
     let output = cli.run(["get", "buckets"]);
@@ -30,39 +24,10 @@ fn test_bucket_lifecycle() {
 }
 
 #[test]
-fn test_bucket_with_invalid_transform_tag_fails() {
-    let cli = TestCli::get();
-    let owner = TestCli::project();
-
-    let new_bucket_name = format!("{}/test-source-{}", owner, Uuid::new_v4());
-
-    let output = cli.run_and_error([
-        "create",
-        "bucket",
-        &new_bucket_name,
-        "--transform-tag",
-        "not-a-valid-transform-tag.0.ABCDEFGH",
-    ]);
-    assert!(
-        output.contains(
-            "422 Unprocessable Entity: The value 'not-a-valid-transform-tag.0.ABCDEFGH' is not a valid transform tag."
-        ),
-        "{}",
-        output,
-    );
-}
-
-#[test]
 fn test_create_without_org_fails() {
     let cli = TestCli::get();
 
-    let output = cli.run_and_error([
-        "create",
-        "bucket",
-        "bucket-name-without-org",
-        "--transform-tag",
-        "generic_simple.0.QR5PW4VZ",
-    ]);
+    let output = cli.run_and_error(["create", "bucket", "bucket-name-without-org"]);
     assert!(
         output.contains("Expected <owner>/<name>, got: bucket-name-without-org"),
         "{}",
@@ -74,13 +39,7 @@ fn test_create_without_org_fails() {
 fn test_create_with_empty_org_fails() {
     let cli = TestCli::get();
 
-    let output = cli.run_and_error([
-        "create",
-        "bucket",
-        "/bucket-name-with-empty-org",
-        "--transform-tag",
-        "generic_simple.0.QR5PW4VZ",
-    ]);
+    let output = cli.run_and_error(["create", "bucket", "/bucket-name-with-empty-org"]);
     assert!(
         output.contains("Expected <owner>/<name>, got: /bucket-name-with-empty-org"),
         "{}",
@@ -92,13 +51,7 @@ fn test_create_with_empty_org_fails() {
 fn test_create_with_empty_bucket_name_fails() {
     let cli = TestCli::get();
 
-    let output = cli.run_and_error([
-        "create",
-        "bucket",
-        "org-without-bucket-name/",
-        "--transform-tag",
-        "generic_simple.0.QR5PW4VZ",
-    ]);
+    let output = cli.run_and_error(["create", "bucket", "org-without-bucket-name/"]);
     assert!(
         output.contains("Expected <owner>/<name>, got: org-without-bucket-name/"),
         "{}",
@@ -110,13 +63,7 @@ fn test_create_with_empty_bucket_name_fails() {
 fn test_create_with_too_many_seperators_fails() {
     let cli = TestCli::get();
 
-    let output = cli.run_and_error([
-        "create",
-        "bucket",
-        "Bucket/Name/with/too/many/seperators/",
-        "--transform-tag",
-        "generic_simple.0.QR5PW4VZ",
-    ]);
+    let output = cli.run_and_error(["create", "bucket", "Bucket/Name/with/too/many/seperators/"]);
     assert!(
         output.contains("Expected <owner>/<name>, got: Bucket/Name/with/too/many/seperators"),
         "{}",